+++ EQM-GQE.SHARED-RESOURCES/Interfaces/IQuestionnaireRepository.cs
@@ -9,7 +9,7 @@ namespace EQM_GQE.SHARED_RESOURCES.Interfaces
 {
     public interface IQuestionnaireRepository
     {
-        Task<long> Add(Questionnaire Questionnaire);
+        Task<int> Add(Questionnaire Questionnaire);
 
         Task<bool> Update(Questionnaire Questionnaire);
